# Navi card subgroup shuffle support for gemm