--- a/src/kernels/level3/xgemm_part3.opencl
+++ b/src/kernels/level3/xgemm_part3.opencl
@@ -31,7 +31,7 @@ INLINE_FUNC int clblast_get_sub_group_local_id() {
   asm volatile("mov.u32 %0, %%laneid;" : "=r"(ret) );
   return ret;
   #elif SUBGROUP_SHUFFLING_GCN == 1
-  return get_local_id(0);
+  return get_sub_group_local_id();
   #endif 
 }
 
@@ -52,20 +52,7 @@ INLINE_FUNC realN clblast_sub_group_shuffle(realN reg, int src) {
     #endif
   return ret;
   #elif SUBGROUP_SHUFFLING_GCN == 1
-#define GCN_SHFL_32(s0, l)  \
-    { \
-		__asm ( \
-		  "ds_bpermute_b32  %[os0], %[ol0], %[os0]\n" \
-          "s_waitcnt lgkmcnt(0)\n" \
-		  : [os0] "=&v" (s0) \
-		  : [ol0] "v" (l), \
-            [os0] "0" (s0)); \
-	}
-
-  realN ret = reg;
-  uint lane_src = (src) << 2;
-  GCN_SHFL_32(ret, lane_src)
-  return ret;
+  return sub_group_shuffle(reg, src);
   #endif
 }
 #endif