--- a/src/utilities/compile.cpp
+++ b/src/utilities/compile.cpp
@@ -78,6 +78,11 @@ std::shared_ptr<Program> CompileFromSource(
     }
   }
 
+  if (device.IsGPU() && device.IsAMD()) {
+    header_string += "#define USE_SUBGROUP_SHUFFLING 1\n";
+    header_string += "#define SUBGROUP_SHUFFLING_GCN 1\n";
+  }
+
   // For Qualcomm devices, specifying the OpenCL kernel attribute reqd_work_group_size reduces performance.
   // This option compiles without the workgroup size requirement and does not affect correctness.
   if (device.IsQualcomm()) {