WIP

Question: how to handle all realN types?