@CNugteren 2023-11-03
> Seems like if SUBGROUP_SIZE the tuner will never satisfy this condition. Don't know how nvidia's 32 subgroup size can work.

You are right that this can be an issue, as said even in the original post of the PR that introduced it (https://github.com/CNugteren/CLBlast/pull/297). However, a bit further down in that thread I do make some analysis (https://github.com/CNugteren/CLBlast/pull/297#issuecomment-406866302) which concludes that there are some combinations that trigger it, although not many, and that they might have to be changed to increase the chances - which we never did I believe. However a user is free to modify the tuner files manually of course.

Regarding the PR itself, apart from my earlier comment (to see if shuffling here is always a good idea), I have 3 more comments before we can merge this in:
1. Could you add a short statement about this to the Changelog file such that it is incorporated in the release notes of the next version? Thanks!
2. Could you also test performance on at least one other AMD Navi GPU if you have access to it? Or perhaps someone else can do that?
3. Could you also run the CLBlast XGEMM tests on your GPU with your changes and with the new best tuning parameters?

Again, thanks a lot for your contribution!