@CNugteren 2023-11-08
>  One thing I don't understand is ...

I re-read it and looked at the code again and I don't understand the comment either. I would rather trust the code than the comment, perhaps I made a typo in the comment and I meant larger instead of smaller, I might have gotten confused by the fact that it says `MDIMC < SUBGROUP_SIZE` but that is to disable it of course.