@tangjinchuan 2024-06-08
[7800xt.txt](https://github.com/user-attachments/files/15748334/7800xt.txt)
Using the latest AMD RoCM 6.1.2 on Ubuntu 24.04 with the repository  ed32af0ea236216468653d92ddcf6f219eebc5dd with AMD shuffle. There is no problem reported.