@fancyIX 2024-06-09
@tangjinchuan Sorry I don't have time to finish this PR recently. To me it's not mature to apply AMD assembly shuffle yet. If you read previous posts in this thread, only certain configuration of matrix size will trigger assembly shuffle. For most cases it's not triggered. I am not sure if shuffle is needed at all.